refactor(cli): split cli_request.go by request/diagnostic family
pkg/cli/cli_request.go had grown to 1328 lines: a grab-bag of the
operational request and diagnostic command handlers (ping/traceroute,
test policy/routing/security-zone, monitor traffic, and the full
request chassis/system/security/dhcp/protocols trees) with their argv
builders and parsing helpers all in one file. This made the file hard
to navigate and review.

Split by command family into cohesive sibling files, keeping
cli_request.go as the dispatch shell. This is pure code motion — every
handler and helper moved verbatim, no behavior change. Because the
files stay in package cli, the unexported symbols remain reachable.

New files:

  cli_request_ping.go      ping / traceroute + diagcmd argv builders
  cli_request_testcmd.go   test policy / routing / security-zone
  monitor_traffic.go       monitor dispatch + monitor traffic (tcpdump)
  cli_request_chassis.go   request chassis cluster failover / data-plane
  cli_request_system.go    request system reboot/halt/zeroize/software/...
  cli_request_security.go  request security ipsec / policies / wireguard

cli_request.go retains the handleRequest dispatcher plus the small
request dhcp / request protocols handlers that do not warrant their own
file. The monitor traffic code lands in monitor_traffic.go to join the
existing sibling monitor.go / monitor_interface.go rather than a
request-prefixed name, since monitor is not a request subcommand. The
test-command file is named cli_request_testcmd.go (not _test.go) so the
Go toolchain does not mistake it for a test file and drop it from the
package build.

The security-sensitive diagcmd/tcpdump argv construction with its "--"
end-of-options separators (#2084, #4524, #4527) moved byte-for-byte;
verified identical against origin/master along with all 31 top-level
declarations.

Validation: go build ./... clean, go vet ./pkg/cli/... shows only the
pre-existing cli.go:503 unreachable-return finding (untouched file),
go test ./pkg/cli/... green, gofmt -l clean on all new and trimmed
files.

Closes #4653
