+++ pkg/cli/cli_request.go
@@ -1,757 +1,27 @@
 package cli
 
-import (
-	"context"
-	"fmt"
-	"net"
-	"os"
-	"os/exec"
-	"strconv"
-	"strings"
-	"time"
-
-	"github.com/psaab/xpf/pkg/cluster"
-	"github.com/psaab/xpf/pkg/config"
-	dpuserspace "github.com/psaab/xpf/pkg/dataplane/userspace"
-	dpformat "github.com/psaab/xpf/pkg/dataplane/userspace/format"
-	"github.com/psaab/xpf/pkg/diagcmd"
-	"github.com/psaab/xpf/pkg/policymatch"
-	"github.com/psaab/xpf/pkg/routing"
-	"github.com/psaab/xpf/pkg/wgkey"
-)
-
-func (c *CLI) handlePing(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("usage: ping <target> [count <N>] [source <IP>] [size <N>] [routing-instance <name>]")
-		return nil
-	}
-
-	target := args[0]
-	count := "5"
-	source := ""
-	size := ""
-	vrfName := ""
-
-	for i := 1; i < len(args)-1; i++ {
-		switch args[i] {
-		case "count":
-			count = args[i+1]
-			i++
-		case "source":
-			source = args[i+1]
-			i++
-		case "size":
-			size = args[i+1]
-			i++
-		case "routing-instance":
-			vrfName = args[i+1]
-			i++
-		}
-	}
-
-	cmdArgs := buildPingArgv(target, count, source, size, vrfName)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
-	defer cancel()
-	c.cmdMu.Lock()
-	c.cmdCancel = cancel
-	c.cmdMu.Unlock()
-	defer func() {
-		c.cmdMu.Lock()
-		c.cmdCancel = nil
-		c.cmdMu.Unlock()
-	}()
-
-	cmd := exec.CommandContext(ctx, cmdArgs[0], cmdArgs[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if ctx.Err() != nil {
-		return nil // cancelled by Ctrl-C or timeout
-	}
-	return err
-}
-
-// buildPingArgv builds the argv for the local CLI ping command. It
-// delegates to the shared diagcmd builder so the VRF-device
-// normalization (apply "vrf-" exactly once, #2143) and the "--"
-// end-of-options separator (#2084) match the REST and gRPC surfaces
-// byte-for-byte. Before #2143 this path prepended "vrf-"
-// unconditionally, turning `routing-instance vrf-red` into the
-// non-existent device `vrf-vrf-red`.
-func buildPingArgv(target, count, source, size, vrfName string) []string {
-	return diagcmd.PingArgv(diagcmd.PingOptions{
-		Target:          target,
-		Count:           count,
-		Source:          source,
-		Size:            size,
-		RoutingInstance: vrfName,
-	})
-}
-
-func (c *CLI) handleTraceroute(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("usage: traceroute <target> [source <IP>] [routing-instance <name>]")
-		return nil
-	}
-
-	target := args[0]
-	source := ""
-	vrfName := ""
-
-	for i := 1; i < len(args)-1; i++ {
-		switch args[i] {
-		case "source":
-			source = args[i+1]
-			i++
-		case "routing-instance":
-			vrfName = args[i+1]
-			i++
-		}
-	}
-
-	cmdArgs := buildTracerouteArgv(target, source, vrfName)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
-	defer cancel()
-	c.cmdMu.Lock()
-	c.cmdCancel = cancel
-	c.cmdMu.Unlock()
-	defer func() {
-		c.cmdMu.Lock()
-		c.cmdCancel = nil
-		c.cmdMu.Unlock()
-	}()
-
-	cmd := exec.CommandContext(ctx, cmdArgs[0], cmdArgs[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if ctx.Err() != nil {
-		return nil // cancelled by Ctrl-C or timeout
-	}
-	return err
-}
-
-// buildTracerouteArgv builds the argv for the local CLI traceroute
-// command. Like buildPingArgv it delegates to the shared diagcmd builder
-// so VRF normalization (#2143) and the "--" separator (#2084) stay
-// identical across the CLI, REST, and gRPC surfaces.
-func buildTracerouteArgv(target, source, vrfName string) []string {
-	return diagcmd.TracerouteArgv(diagcmd.TracerouteOptions{
-		Target:          target,
-		Source:          source,
-		RoutingInstance: vrfName,
-	})
-}
-
-// handleTest dispatches test sub-commands (policy, routing, security-zone).
-func (c *CLI) handleTest(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("test: specify a test command")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["test"].Children))
-		return nil
-	}
-
-	resolved, err := resolveCommand(args[0], keysFromTree(operationalTree["test"].Children))
-	if err != nil {
-		return err
-	}
-
-	switch resolved {
-	case "policy":
-		return c.testPolicy(args[1:])
-	case "routing":
-		return c.testRouting(args[1:])
-	case "security-zone":
-		return c.testSecurityZone(args[1:])
-	default:
-		return fmt.Errorf("unknown test command: %s", resolved)
-	}
-}
-
-// testPolicy performs a 5-tuple policy lookup similar to Junos "test policy".
-func (c *CLI) testPolicy(args []string) error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("No active configuration")
-		return nil
-	}
-
-	// #3696: parse through the single strict SSOT selector parser shared by all
-	// four CLI surfaces + the gRPC test-policy bridge. A value-taking selector
-	// with no value, an unknown selector token, an explicit-empty typed value,
-	// and a malformed IP / port / protocol / icmp value are HARD ERRORS instead
-	// of silently degrading to the wildcard — the simulator answers the query
-	// the operator typed, not a broader one. Per-value validation (#3116 /
-	// #3108 / #3284 / #1711) now lives in the parser.
-	sel, err := policymatch.ParseSelectorArgs(args)
-	if err != nil {
-		return err
-	}
-	fromZone, toZone, srcIP, dstIP, proto := sel.FromZone, sel.ToZone, sel.SrcIP, sel.DstIP, sel.Protocol
-	srcPort, dstPort := sel.SrcPort, sel.DstPort
-	icmpType, icmpCode := sel.ICMPType, sel.ICMPCode
-
-	if fromZone == "" || toZone == "" {
-		// #3628: shared SSOT selector list (policymatch) so the advertised
-		// selectors match what the parser accepts, including icmp-type/icmp-code
-		// and protocol by name or number.
-		fmt.Println(policymatch.TestPolicyUsage)
-		return nil
-	}
-
-	parsedSrc := net.ParseIP(srcIP)
-	parsedDst := net.ParseIP(dstIP)
-
-	// #3042: delegate to the single shared simulator (exact zone-pair ->
-	// wildcard-zone tiers (#3090) -> scoped global (#3148) -> default-policy).
-	// The pre-#3042 loop hard-coded "Default deny" (ignoring default-policy
-	// permit-all) and used a narrow address/app matcher that missed predefined
-	// apps, nested application-sets, literal CIDRs, any-ipv4/any-ipv6, and
-	// source/destination exclusion.
-	// #3105: pass the live dynamic-address feed-prefix overlay so a feed-backed
-	// address-name resolves to its live CIDRs on-box, matching the REST/gRPC
-	// simulators and the AF_XDP helper. Nil (CLI outside the daemon) keeps the
-	// pre-#3105 static-only behavior.
-	res := policymatch.Match(cfg, policymatch.Query{
-		FromZone:    fromZone,
-		ToZone:      toZone,
-		SrcIP:       parsedSrc,
-		DstIP:       parsedDst,
-		Protocol:    proto,
-		SrcPort:     srcPort,
-		DstPort:     dstPort,
-		ICMPType:    icmpType,
-		ICMPCode:    icmpCode,
-		FeedOverlay: c.feedOverlay(),
-		// #3104: skip scheduler-inactive policies like the runtime does, so the
-		// simulator falls through to the next active rule / default-policy.
-		PolicyInactiveFn: c.policyInactiveFn(),
-	})
-	if res.ContentRejected {
-		// #3727: the dataplane fails this config closed (unexpandable
-		// application-set) and enforces none of its policies — do NOT print a
-		// fabricated permit/deny/default verdict.
-		fmt.Printf("Policy content rejected (no verdict enforced for %s -> %s)\n", fromZone, toZone)
-		fmt.Printf("  %s\n", policymatch.ContentRejectedShowLine)
-		for _, reason := range res.ContentRejectionReasons {
-			fmt.Printf("    %s\n", reason)
-		}
-		return nil
-	}
-	if res.HostInboundUnmatched {
-		// #3285: host-bound traffic — no transit global/default fallback.
-		fmt.Printf("No matching to-zone junos-host policy for %s -> junos-host\n", fromZone)
-		fmt.Printf("  %s\n", policymatch.HostInboundShowLine)
-		return nil
-	}
-	// #4373 (E4/H2/H7): a multicast/broadcast/unspecified/loopback destination is
-	// dropped at route lookup before policy runs — surface the advisory so the
-	// `test policy` verdict below is not read as real forwarding.
-	if note := res.RouteDropNote(); note != "" {
-		fmt.Printf("  %s\n", note)
-	}
-	if !res.Matched {
-		fmt.Printf("Default %s (no matching policy for %s -> %s)\n",
-			policymatch.ActionString(res.Action), fromZone, toZone)
-		return nil
-	}
-	if res.Global {
-		fmt.Printf("Policy match (global):\n")
-		fmt.Printf("  Policy:    %s\n", res.PolicyName)
-		printPolicyMatchIdentity(res)
-		fmt.Printf("  Action:    %s\n", policymatch.ActionString(res.Action))
-		return nil
-	}
-	fmt.Printf("Policy match:\n")
-	fmt.Printf("  From zone: %s\n  To zone:   %s\n", fromZone, toZone)
-	fmt.Printf("  Policy:    %s\n", res.PolicyName)
-	printPolicyMatchIdentity(res)
-	fmt.Printf("  Action:    %s\n", policymatch.ActionString(res.Action))
-	if srcIP != "" {
-		fmt.Printf("  Source:    %s -> ", srcIP)
-	} else {
-		fmt.Printf("  Source:    any -> ")
-	}
-	if dstIP != "" {
-		fmt.Printf("%s", dstIP)
-	} else {
-		fmt.Printf("any")
-	}
-	if dstPort > 0 {
-		fmt.Printf(":%d", dstPort)
-	}
-	// #4497 (avo-001 F3): echo the queried ICMP/ICMPv6 type and code alongside
-	// the protocol so the `test policy` verdict names the FULL tuple the
-	// simulator matched, not just the protocol. A
-	// `protocol icmp icmp-type 8 icmp-code 0` query is answered against the
-	// declared type/code (junos-ping = type 8, #3284); dropping the type/code
-	// from the echo hid WHICH ICMP packet was tested. A non-ICMP query (no
-	// type/code) prints the bare `[proto]` exactly as before.
-	if tail := formatQueryProtoTail(proto, icmpType, icmpCode); tail != "" {
-		fmt.Printf(" %s", tail)
-	}
-	fmt.Println()
-	return nil
-}
-
-// formatQueryProtoTail renders the trailing `[proto ...]` annotation for the
-// `test policy` query echo (#4497, avo-001 F3). It surfaces the queried
-// ICMP/ICMPv6 type and code so the operator reads the exact tuple the simulator
-// matched, not just the protocol name. A nil type/code is omitted; an empty
-// protocol with no type/code yields "" (nothing to annotate).
-func formatQueryProtoTail(proto string, icmpType, icmpCode *uint8) string {
-	if proto == "" && icmpType == nil && icmpCode == nil {
-		return ""
-	}
-	var b strings.Builder
-	b.WriteByte('[')
-	sep := ""
-	if proto != "" {
-		b.WriteString(proto)
-		sep = " "
-	}
-	if icmpType != nil {
-		fmt.Fprintf(&b, "%stype %d", sep, *icmpType)
-		sep = " "
-	}
-	if icmpCode != nil {
-		fmt.Fprintf(&b, "%scode %d", sep, *icmpCode)
-	}
-	b.WriteByte(']')
-	return b.String()
-}
-
-// printPolicyMatchIdentity renders the shared PolicyID / RuleID / scope /
-// description block for the local `test policy` simulator (#3674). Before this,
-// the request-path output printed only the policy name + action (and for a
-// global match, nothing but name + action), while `show security match-policies`
-// over the SAME policymatch.Result already printed these identity fields. That
-// made `test policy` the poorest of the match-policies surfaces: an operator
-// could not correlate a verdict with the RT_FLOW / session-table policy ID or
-// tell whether a global vs zone-pair rule fired.
+// cli_request.go is the dispatch shell for the operational request /
+// diagnostic command families that were historically one 1300-line grab-bag
+// (#4653). It keeps only the top-level `request` dispatcher plus the small
+// `request dhcp` / `request protocols` handlers that do not warrant their own
+// file. The per-family handlers live in sibling files in this package (same
+// package, so unexported symbols stay reachable):
 //
-// All fields come straight off policymatch.Result (populated by the #3667
-// RuntimePolicyIDs SSOT), and the scope line reuses the shared matchScopeZone
-// helper (#3331), so this surface stays in lock-step with showMatchPolicies with
-// no duplicated logic. The 2-space aligned label style matches the surrounding
-// `test policy` output rather than the show path's 4-space sub-indent.
-func printPolicyMatchIdentity(res policymatch.Result) {
-	fmt.Printf("  Policy ID: %d\n", res.PolicyID)
-	// #3668: the stable rule identity joins a hit to the inventory / logs /
-	// tests even after a reorder shifts the numeric Policy ID.
-	if res.RuleID != "" {
-		fmt.Printf("  Rule ID:   %s\n", res.RuleID)
-	}
-	if res.Global {
-		fmt.Printf("  Scope:     global (match from-zone: %s, to-zone: %s)\n",
-			matchScopeZone(res.FromZone), matchScopeZone(res.ToZone))
-	} else {
-		fmt.Printf("  Scope:     zone-pair (from-zone: %s, to-zone: %s)\n",
-			res.FromZone, res.ToZone)
-	}
-	if res.Description != "" {
-		fmt.Printf("  Description: %s\n", res.Description)
-	}
-}
+//	cli_request_ping.go     ping / traceroute (+ diagcmd argv builders)
+//	cli_request_testcmd.go  test policy / routing / security-zone
+//	monitor_traffic.go      monitor + monitor traffic (tcpdump argv builder)
+//	cli_request_chassis.go  request chassis cluster failover / data-plane
+//	cli_request_system.go   request system reboot/halt/zeroize/software/...
+//	cli_request_security.go request security ipsec / policies / wireguard
+//
+// The split is pure code motion — no behavior change. The security-sensitive
+// diagcmd/tcpdump argv construction with its `--` end-of-options separators
+// (#4527 / #2084 / #4524) moved verbatim.
 
-// testRouting looks up a destination in the routing table.
-func (c *CLI) testRouting(args []string) error {
-	if c.routing == nil {
-		fmt.Println("Routing manager not available")
-		return nil
-	}
-
-	var dest, instance string
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "destination":
-			if i+1 < len(args) {
-				i++
-				dest = args[i]
-			}
-		case "instance":
-			if i+1 < len(args) {
-				i++
-				instance = args[i]
-			}
-		}
-	}
-
-	if dest == "" {
-		fmt.Println("usage: test routing destination <ip-or-prefix> [instance <name>]")
-		return nil
-	}
-
-	var entries []routing.RouteEntry
-	var err error
-	if instance != "" {
-		entries, err = c.routing.GetVRFRoutes(instance)
-	} else {
-		entries, err = c.routing.GetRoutes()
-	}
-	if err != nil {
-		return fmt.Errorf("get routes: %w", err)
-	}
-
-	// Normalize dest to CIDR for matching
-	filterCIDR := dest
-	if !strings.Contains(filterCIDR, "/") {
-		if strings.Contains(filterCIDR, ":") {
-			filterCIDR += "/128"
-		} else {
-			filterCIDR += "/32"
-		}
-	}
-	filterIP, _, filterErr := net.ParseCIDR(filterCIDR)
-	if filterErr != nil {
-		filterIP = net.ParseIP(dest)
-	}
-
-	// Find the best (longest prefix) match
-	var best *routing.RouteEntry
-	bestLen := -1
-	for i := range entries {
-		_, rNet, err := net.ParseCIDR(entries[i].Destination)
-		if err != nil {
-			continue
-		}
-		if filterIP != nil && rNet.Contains(filterIP) {
-			ones, _ := rNet.Mask.Size()
-			if ones > bestLen {
-				bestLen = ones
-				best = &entries[i]
-			}
-		}
-	}
-
-	if instance != "" {
-		fmt.Printf("Routing lookup in instance %s for %s:\n", instance, dest)
-	} else {
-		fmt.Printf("Routing lookup for %s:\n", dest)
-	}
-	if best == nil {
-		fmt.Println("  No matching route found")
-	} else {
-		fmt.Printf("  Destination: %s\n", best.Destination)
-		fmt.Printf("  Next-hop:    %s\n", best.NextHop)
-		fmt.Printf("  Interface:   %s\n", best.Interface)
-		fmt.Printf("  Protocol:    %s\n", best.Protocol)
-		fmt.Printf("  Preference:  %d\n", best.Preference)
-	}
-	return nil
-}
-
-// testSecurityZone looks up which zone an interface belongs to.
-func (c *CLI) testSecurityZone(args []string) error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("No active configuration")
-		return nil
-	}
-
-	var ifName string
-	for i := 0; i < len(args); i++ {
-		if args[i] == "interface" && i+1 < len(args) {
-			i++
-			ifName = args[i]
-		}
-	}
-
-	if ifName == "" {
-		fmt.Println("usage: test security-zone interface <name>")
-		return nil
-	}
-
-	for zoneName, zone := range cfg.Security.Zones {
-		if zone == nil { // #3493: tolerant/HA-sync path may carry a nil zone value
-			continue
-		}
-		for _, iface := range zone.Interfaces {
-			if iface == ifName {
-				fmt.Printf("Interface %s belongs to zone: %s\n", ifName, zoneName)
-				if zone.Description != "" {
-					fmt.Printf("  Description: %s\n", zone.Description)
-				}
-				if zone.ScreenProfile != "" {
-					fmt.Printf("  Screen:      %s\n", zone.ScreenProfile)
-				}
-				// #3654 (H06/M03): this DIAGNOSTIC is supposed to explain
-				// per-interface admission, so report the EFFECTIVE (zone UNION
-				// interface) set for THIS interface, flag an interface-local
-				// override, and print an explicit default-deny posture line.
-				// #3682: also flag when THIS interface is a management /
-				// cluster-control lifeline excluded from host-inbound deny.
-				lifeline := config.HostInboundLifelineInterface(
-					ifName, config.HostInboundLifelineSet(cfg))
-				for _, line := range zone.RenderInterfaceHostInbound(ifName, lifeline,
-					config.HostInboundLabels{
-						Indent:         "  ",
-						Sep:            ", ",
-						ServicesLabel:  "Host-inbound services",
-						ProtocolsLabel: "Host-inbound protocols",
-					}) {
-					fmt.Println(line)
-				}
-				return nil
-			}
-		}
-	}
-
-	fmt.Printf("Interface %s is not assigned to any security zone\n", ifName)
-	return nil
-}
-
-// handleMonitor dispatches monitor sub-commands.
-func (c *CLI) handleMonitor(args []string) error {
-	monTree := operationalTree["monitor"].Children
-	if len(args) == 0 {
-		fmt.Println("monitor:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(monTree))
-		return nil
-	}
-
-	resolved, err := resolveCommand(args[0], keysFromTree(monTree))
-	if err != nil {
-		return err
-	}
-
-	switch resolved {
-	case "traffic":
-		return c.handleMonitorTraffic(args[1:])
-	case "interface":
-		return c.handleMonitorInterface(args[1:])
-	case "security":
-		return c.handleMonitorSecurity(args[1:])
-	default:
-		return fmt.Errorf("unknown monitor target: %s", resolved)
-	}
-}
-
-// monitorTrafficKeywords are the option keywords the monitor traffic
-// grammar recognizes. They terminate a greedy `matching <filter>`
-// expression so a multi-token filter never swallows a following option
-// (e.g. `matching tcp port 80 count 20`).
-var monitorTrafficKeywords = map[string]bool{
-	"interface": true,
-	"matching":  true,
-	"count":     true,
-}
-
-// parseMonitorTrafficArgs parses the `monitor traffic ...` argument
-// vector into the interface name, the pcap/BPF filter expression, and the
-// packet count. The CLI tokenizer (strings.Fields) splits on whitespace
-// and does NOT honor shell quoting, so a multi-token filter typed as
-// `matching "tcp port 80"` arrives as the separate tokens `"tcp`, `port`,
-// `80"`. The `matching` clause therefore consumes EVERY following token up
-// to the next recognized option keyword (none of which are pcap
-// primitives), joins them into one filter expression, and strips any
-// surrounding quotes the operator typed. Reading only the first token
-// truncated `tcp port 80` down to `tcp` and silently captured far more
-// traffic than requested (#4005).
-func parseMonitorTrafficArgs(args []string) (iface, filter, count string, err error) {
-	count = "0" // 0 = unlimited
-
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "interface":
-			// The value after `interface` must be a real interface name,
-			// not a missing value or another grammar keyword. Without this
-			// guard `interface matching tcp port 80` consumed `matching` as
-			// the interface name → tcpdump "no such device" or an
-			// unfiltered capture; a trailing bare `interface` consumed
-			// nothing and fell through to the usage message (#4540).
-			if i+1 >= len(args) || monitorTrafficKeywords[args[i+1]] {
-				return "", "", "", fmt.Errorf("monitor traffic: 'interface' requires a value")
-			}
-			i++
-			iface = args[i]
-		case "matching":
-			rest := make([]string, 0, len(args)-i-1)
-			for j := i + 1; j < len(args); j++ {
-				if monitorTrafficKeywords[args[j]] {
-					break
-				}
-				rest = append(rest, args[j])
-			}
-			i += len(rest)
-			filter = stripSurroundingQuotes(strings.Join(rest, " "))
-		case "count":
-			// `count` needs a numeric value. A missing value (a trailing
-			// bare `count`, or a following grammar keyword) previously left
-			// count at "0" = unlimited, silently ignoring the operator's
-			// intent to bound the capture; a non-numeric value reached
-			// tcpdump's `-c` and produced an opaque error (#4540).
-			if i+1 >= len(args) || monitorTrafficKeywords[args[i+1]] {
-				return "", "", "", fmt.Errorf("monitor traffic: 'count' requires a value")
-			}
-			i++
-			count = args[i]
-			n, cerr := strconv.Atoi(count)
-			if cerr != nil {
-				return "", "", "", fmt.Errorf("monitor traffic: 'count' requires a numeric value, got %q", count)
-			}
-			// Bound the count like the sibling `monitor security packet-drop`
-			// (1..8192, monitor.go). 0 stays the explicit "unlimited" mode
-			// (buildMonitorTrafficArgv omits `-c` when count=="0"); a negative
-			// previously reached tcpdump's `-c` as an opaque "invalid packet
-			// count" error, and an unbounded huge value is redundant with the
-			// already-supported 0=unlimited mode. Reject both CLI-side so the
-			// numeric leaf matches the project's sibling-command bounding.
-			if n < 0 || n > 8192 {
-				return "", "", "", fmt.Errorf("monitor traffic: 'count' must be 0 (unlimited) or 1..8192, got %q", count)
-			}
-		}
-	}
-	return iface, filter, count, nil
-}
-
-// stripSurroundingQuotes removes one layer of matching leading/trailing
-// single or double quotes. The CLI tokenizer does not honor shell quoting,
-// so an operator who wraps a multi-token pcap filter in quotes
-// (`matching "tcp port 80"`) leaves literal quote characters on the first
-// and last tokens; tcpdump/libpcap would reject those literal quotes, so
-// peel one balanced layer here.
-func stripSurroundingQuotes(s string) string {
-	if len(s) >= 2 {
-		q := s[0]
-		if (q == '"' || q == '\'') && s[len(s)-1] == q {
-			return s[1 : len(s)-1]
-		}
-	}
-	return s
-}
-
-// buildMonitorTrafficArgv assembles the tcpdump argv for a live capture.
-// The full filter expression is passed as trailing arguments exactly as
-// tcpdump/libpcap expects a filter (all tokens, verbatim), but only AFTER
-// an explicit "--" end-of-options separator so no filter token can be
-// interpreted as a tcpdump option.
-func buildMonitorTrafficArgv(iface, filter, count string) []string {
-	cmdArgs := []string{"tcpdump", "-i", iface, "-n", "-l"}
-	if count != "0" {
-		cmdArgs = append(cmdArgs, "-c", count)
-	}
-	if filter != "" {
-		// Insert a "--" end-of-options separator before the operator-supplied
-		// filter (#4524), mirroring the diagcmd ping/traceroute #2084
-		// treatment. The `matching <filter>` clause greedily absorbs every
-		// token up to the next keyword, so without the separator a filter
-		// like `-w /etc/cron.d/x` or `-z <cmd>` reaches tcpdump as an OPTION
-		// under glibc getopt argv permutation — arbitrary root file-write
-		// (-w) or post-rotate command execution (-z) past the capture-only
-		// RBAC boundary. After "--", getopt stops scanning for options and
-		// tcpdump treats every remaining token as part of the pcap filter
-		// EXPRESSION, so an injected `-w`/`-z` becomes a (harmless) filter
-		// operand that libpcap rejects at compile time rather than an option.
-		//
-		// tcpdump accepts the filter expression as separate argv tokens;
-		// splitting on whitespace keeps a multi-token filter intact and
-		// matches how tcpdump joins its own trailing filter arguments.
-		cmdArgs = append(cmdArgs, "--")
-		cmdArgs = append(cmdArgs, strings.Fields(filter)...)
-	}
-	return cmdArgs
-}
-
-// monitorFilterOptionToken reports whether tok would be interpreted by
-// tcpdump/getopt as an option flag rather than as a pcap filter primitive.
-// A legitimate pcap filter expression never begins a term with an option
-// flag (`-w`, `-z`, `-r`, `--postrotate-command`, ...): filter primitives
-// are keywords (host/port/tcp/udp/net/...), addresses, numbers, boolean
-// operators (and/or/not) and parentheses. Such an option-looking token can
-// therefore only be an attempt to smuggle a tcpdump option past the
-// capture-only contract (#4524). A bare "-" (getopt's stdin sentinel, not
-// an option) is allowed so an arithmetic subtraction term is never falsely
-// rejected. This is defense-in-depth: the "--" separator in
-// buildMonitorTrafficArgv already neutralizes the option, but rejecting the
-// token here gives the operator a clear error instead of an opaque libpcap
-// compile failure.
-func monitorFilterOptionToken(tok string) bool {
-	// #4556 N-01: a mismatched-quote wrapper (e.g. `'-w ...` / `"-z ...`)
-	// can leave a literal leading quote on the token, so tok[0] is `'`/`"`
-	// instead of `-` and a smuggled option would slip past the tok[0]=='-'
-	// check. Peel one leading quote before testing. The "--" separator in
-	// buildMonitorTrafficArgv already neutralizes any such option downstream
-	// (#4527), so this only closes a defense-in-depth validator gap — but it
-	// keeps the operator error clear rather than opaque. A legitimate pcap
-	// filter primitive never begins with a quote, so peeling one is a no-op
-	// for normal tokens (host/port/tcp/... and a bare "-").
-	if len(tok) > 0 && (tok[0] == '\'' || tok[0] == '"') {
-		tok = tok[1:]
-	}
-	return len(tok) > 1 && tok[0] == '-'
-}
-
-// validateMonitorFilter rejects a pcap filter that contains an
-// option-looking token (see monitorFilterOptionToken). Returns nil for an
-// empty filter or any legitimate pcap expression.
-func validateMonitorFilter(filter string) error {
-	for _, tok := range strings.Fields(filter) {
-		if monitorFilterOptionToken(tok) {
-			return fmt.Errorf("invalid filter token %q: a capture filter cannot contain a tcpdump option flag", tok)
-		}
-	}
-	return nil
-}
-
-// handleMonitorTraffic wraps tcpdump for live packet capture.
-func (c *CLI) handleMonitorTraffic(args []string) error {
-	iface, filter, count, err := parseMonitorTrafficArgs(args)
-	if err != nil {
-		return err
-	}
-
-	if iface == "" {
-		fmt.Println("usage: monitor traffic interface <name> [matching <filter>] [count <N>]")
-		return nil
-	}
-
-	// Defense-in-depth (#4524): reject a filter that smuggles a tcpdump
-	// option (`-w`, `-z`, ...) before it can reach the argv. The "--"
-	// separator in buildMonitorTrafficArgv already neutralizes it, but a
-	// clear rejection beats an opaque libpcap syntax error.
-	if err := validateMonitorFilter(filter); err != nil {
-		return err
-	}
-
-	// Resolve fabric IPVLAN overlays to physical parent (#136).
-	origName := iface
-	iface = resolveFabricParent(iface)
-
-	// Warn about XDP redirect visibility on fabric interfaces (#138).
-	if strings.HasPrefix(origName, "fab") || strings.HasPrefix(origName, "em") {
-		fmt.Println("WARNING: XDP-redirected packets bypass AF_PACKET and will not appear in tcpdump.")
-		fmt.Println("For fabric redirect telemetry, use: show chassis cluster fabric statistics")
-		fmt.Println()
-	}
-
-	cmdArgs := buildMonitorTrafficArgv(iface, filter, count)
-
-	fmt.Printf("Monitoring traffic on %s (Ctrl+C to stop)...\n", iface)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	c.cmdMu.Lock()
-	c.cmdCancel = cancel
-	c.cmdMu.Unlock()
-	defer func() {
-		c.cmdMu.Lock()
-		c.cmdCancel = nil
-		c.cmdMu.Unlock()
-	}()
-
-	cmd := exec.CommandContext(ctx, cmdArgs[0], cmdArgs[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err = cmd.Run()
-	if ctx.Err() != nil {
-		fmt.Println() // newline after ^C
-		return nil
-	}
-	return err
-}
+import (
+	"fmt"
+	"os"
+)
 
 func (c *CLI) handleRequest(args []string) error {
 	if len(args) == 0 {
@@ -774,210 +44,6 @@ func (c *CLI) handleRequest(args []string) error {
 	default:
 		return fmt.Errorf("unknown request target: %s", args[0])
 	}
-}
-
-func (c *CLI) handleRequestChassis(args []string) error {
-	if len(args) == 0 || args[0] != "cluster" {
-		fmt.Println("request chassis:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children))
-		return nil
-	}
-	args = args[1:] // consume "cluster"
-	if len(args) == 0 {
-		fmt.Println("request chassis cluster:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children))
-		return nil
-	}
-	switch args[0] {
-	case "failover":
-		return c.handleRequestChassisClusterFailover(args[1:])
-	case "data-plane":
-		return c.handleRequestChassisClusterDataPlane(args[1:])
-	default:
-		fmt.Println("request chassis cluster:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children))
-		return nil
-	}
-}
-
-func (c *CLI) handleRequestChassisClusterFailover(args []string) error {
-	if c.cluster == nil {
-		return fmt.Errorf("cluster not configured")
-	}
-	// "request chassis cluster failover reset redundancy-group <N>"
-	if len(args) >= 1 && args[0] == "reset" {
-		if len(args) < 3 || args[1] != "redundancy-group" {
-			return fmt.Errorf("usage: request chassis cluster failover reset redundancy-group <N>")
-		}
-		rgID, err := strconv.Atoi(args[2])
-		if err != nil {
-			return fmt.Errorf("invalid redundancy-group ID: %s", args[2])
-		}
-		if err := c.cluster.ResetFailover(rgID); err != nil {
-			return err
-		}
-		fmt.Printf("Failover reset for redundancy group %d\n", rgID)
-		return nil
-	}
-
-	// "request chassis cluster failover data node <N>"
-	if len(args) >= 3 && args[0] == "data" && args[1] == "node" {
-		targetNode, err := strconv.Atoi(args[2])
-		if err != nil {
-			return fmt.Errorf("invalid node ID: %s", args[2])
-		}
-		if !cluster.IsSupportedClusterNodeID(targetNode) {
-			return fmt.Errorf("unsupported cluster failover target node %d", targetNode)
-		}
-		localNode := c.cluster.NodeID()
-		if targetNode != localNode {
-			message, err := c.requestPeerSystemAction(
-				context.Background(),
-				fmt.Sprintf("cluster-failover-data:node%d", targetNode),
-			)
-			if err != nil {
-				return err
-			}
-			fmt.Println(message)
-			return nil
-		}
-
-		dataRGs := c.cluster.DataGroupIDs()
-		if len(dataRGs) == 0 {
-			return fmt.Errorf("no data redundancy groups configured")
-		}
-		moveRGs := make([]int, 0, len(dataRGs))
-		for _, rgID := range dataRGs {
-			if !c.cluster.IsLocalPrimary(rgID) {
-				moveRGs = append(moveRGs, rgID)
-			}
-		}
-		if len(moveRGs) == 0 {
-			fmt.Printf("All data redundancy groups are already primary on node %d\n", targetNode)
-			return nil
-		}
-		if len(moveRGs) == 1 {
-			if err := c.cluster.RequestPeerFailover(moveRGs[0]); err != nil {
-				return err
-			}
-		} else {
-			if err := c.cluster.RequestPeerFailoverBatch(moveRGs); err != nil {
-				return err
-			}
-		}
-		fmt.Printf("Manual failover completed for data redundancy groups %v (transfer committed)\n", moveRGs)
-		return nil
-	}
-
-	// "request chassis cluster failover redundancy-group <N> [node <N>]"
-	if len(args) >= 2 && args[0] == "redundancy-group" {
-		rgID, err := strconv.Atoi(args[1])
-		if err != nil {
-			return fmt.Errorf("invalid redundancy-group ID: %s", args[1])
-		}
-
-		// If "node <N>" is specified, route to the correct node.
-		if len(args) >= 4 && args[2] == "node" {
-			targetNode, err := strconv.Atoi(args[3])
-			if err != nil {
-				return fmt.Errorf("invalid node ID: %s", args[3])
-			}
-			localNode := c.cluster.NodeID()
-			if targetNode == localNode {
-				if err := c.cluster.RequestPeerFailover(rgID); err != nil {
-					return err
-				}
-				fmt.Printf("Manual failover completed for redundancy group %d (transfer committed)\n", rgID)
-				return nil
-			}
-			message, err := c.requestPeerSystemAction(
-				context.Background(),
-				fmt.Sprintf("cluster-failover:%d:node%d", rgID, targetNode),
-			)
-			if err != nil {
-				return err
-			}
-			fmt.Println(message)
-			return nil
-		}
-
-		if err := c.cluster.ManualFailover(rgID); err != nil {
-			return err
-		}
-		fmt.Printf("Manual failover triggered for redundancy group %d\n", rgID)
-		return nil
-	}
-
-	return fmt.Errorf("usage: request chassis cluster failover {redundancy-group <N> [node <N>] | data node <N>}")
-}
-
-func (c *CLI) handleRequestChassisClusterDataPlane(args []string) error {
-	if len(args) == 0 || args[0] != "userspace" {
-		fmt.Println("request chassis cluster data-plane:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children["data-plane"].Children))
-		return nil
-	}
-	provider, err := c.userspaceDataplaneControl()
-	if err != nil {
-		return err
-	}
-	args = args[1:]
-
-	var status dpuserspace.ProcessStatus
-	switch {
-	case len(args) > 0 && args[0] == "inject-packet":
-		slot, mode, extra, err := dpuserspace.ParseInjectPacketCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.Status()
-		if err != nil {
-			return err
-		}
-		req, err := dpuserspace.BuildInjectPacketRequest(slot, mode, extra, status)
-		if err != nil {
-			return err
-		}
-		status, err = provider.InjectPacket(req)
-		if err != nil {
-			return err
-		}
-	case len(args) > 0 && args[0] == "forwarding":
-		armed, err := dpuserspace.ParseForwardingCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.SetForwardingArmed(armed)
-		if err != nil {
-			return err
-		}
-	case len(args) > 0 && args[0] == "queue":
-		queueID, registered, armed, err := dpuserspace.ParseQueueCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.SetQueueState(queueID, registered, armed)
-		if err != nil {
-			return err
-		}
-	case len(args) > 0 && args[0] == "binding":
-		slot, registered, armed, err := dpuserspace.ParseBindingCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.SetBindingState(slot, registered, armed)
-		if err != nil {
-			return err
-		}
-	default:
-		fmt.Println("request chassis cluster data-plane userspace:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children["data-plane"].Children["userspace"].Children))
-		return nil
-	}
-	fmt.Print(dpformat.FormatStatusSummary(status))
-	fmt.Println()
-	fmt.Print(dpformat.FormatBindings(status))
-	return nil
 }
 
 func (c *CLI) handleRequestDHCP(args []string) error {
@@ -1042,287 +108,4 @@ func (c *CLI) handleRequestProtocols(args []string) error {
 	default:
 		return fmt.Errorf("unknown request protocols target: %s", args[0])
 	}
-}
-
-func (c *CLI) handleRequestSystem(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children))
-		return nil
-	}
-
-	switch args[0] {
-	case "reboot":
-		fmt.Print("Reboot the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Reboot cancelled")
-			return nil
-		}
-		fmt.Println("System going down for reboot NOW!")
-		cmd := exec.Command("systemctl", "reboot")
-		return cmd.Run()
-
-	case "halt":
-		fmt.Print("Halt the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Halt cancelled")
-			return nil
-		}
-		fmt.Println("System halting NOW!")
-		cmd := exec.Command("systemctl", "halt")
-		return cmd.Run()
-
-	case "power-off":
-		fmt.Print("Power off the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Power-off cancelled")
-			return nil
-		}
-		fmt.Println("System powering off NOW!")
-		cmd := exec.Command("systemctl", "poweroff")
-		return cmd.Run()
-
-	case "zeroize":
-		fmt.Println("WARNING: This will erase all configuration and return to factory defaults.")
-		fmt.Print("Zeroize the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Zeroize cancelled")
-			return nil
-		}
-
-		// Remove active and candidate configs, rollback history
-		configDir := "/etc/xpf"
-		files, _ := os.ReadDir(configDir)
-		for _, f := range files {
-			if strings.HasSuffix(f.Name(), ".conf") || strings.HasPrefix(f.Name(), "rollback") {
-				os.Remove(configDir + "/" + f.Name())
-			}
-		}
-
-		// Remove BPF pins
-		os.RemoveAll("/sys/fs/bpf/xpf")
-
-		// Remove managed networkd files
-		ndFiles, _ := os.ReadDir("/etc/systemd/network")
-		for _, f := range ndFiles {
-			if strings.HasPrefix(f.Name(), "10-xpf-") {
-				os.Remove("/etc/systemd/network/" + f.Name())
-			}
-		}
-
-		// Remove FRR managed section
-		exec.Command("systemctl", "stop", "xpfd").Run()
-
-		fmt.Println("System zeroized. Configuration erased.")
-		fmt.Println("Reboot to complete factory reset.")
-		return nil
-
-	case "configuration":
-		return c.handleRequestSystemConfiguration(args[1:])
-
-	case "software":
-		return c.handleRequestSystemSoftware(args[1:])
-
-	case "dynamic-dns":
-		return c.handleRequestSystemDynamicDNS(args[1:])
-
-	default:
-		return fmt.Errorf("unknown request system command: %s", args[0])
-	}
-}
-
-// handleRequestSystemDynamicDNS implements `request system dynamic-dns
-// update|check` (#3276): an operator force-now / check-now verb that triggers an
-// immediate DDNS publish out-of-band of the poll cycle. `update` re-asserts
-// every owned record now (force); `check` re-observes and publishes only changed
-// records. Both honor the per-RG owner gate — on a node that masters no RG the
-// daemon returns a clear "not the active node" message and takes no action.
-func (c *CLI) handleRequestSystemDynamicDNS(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system dynamic-dns:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["dynamic-dns"].Children))
-		return nil
-	}
-	if c.surfaceADDNSForceFn == nil {
-		return fmt.Errorf("dynamic-dns: DDNS engine not running")
-	}
-	switch args[0] {
-	case "update":
-		_, msg := c.surfaceADDNSForceFn(true)
-		fmt.Println(msg)
-		return nil
-	case "check":
-		_, msg := c.surfaceADDNSForceFn(false)
-		fmt.Println(msg)
-		return nil
-	default:
-		return fmt.Errorf("unknown request system dynamic-dns command: %s", args[0])
-	}
-}
-
-func (c *CLI) handleRequestSystemSoftware(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system software:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["software"].Children))
-		return nil
-	}
-
-	if args[0] != "in-service-upgrade" {
-		return fmt.Errorf("unknown request system software command: %s", args[0])
-	}
-
-	if c.cluster == nil {
-		fmt.Println("Cluster not configured")
-		return nil
-	}
-
-	fmt.Println("WARNING: This will force this node to secondary for all redundancy groups.")
-	fmt.Print("Proceed with in-service upgrade? [yes,no] (no) ")
-	c.rl.SetPrompt("")
-	line, err := c.rl.Readline()
-	c.rl.SetPrompt(c.operationalPrompt())
-	if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-		fmt.Println("ISSU cancelled")
-		return nil
-	}
-
-	if err := c.cluster.ForceSecondary(); err != nil {
-		return fmt.Errorf("ISSU: %v", err)
-	}
-
-	fmt.Println("Node is now secondary for all redundancy groups.")
-	fmt.Println("Traffic has been drained to peer.")
-	fmt.Println("You may now replace the binary and restart the service:")
-	fmt.Println("  systemctl stop xpfd && <replace binary> && systemctl start xpfd")
-	return nil
-}
-
-func (c *CLI) handleRequestSystemConfiguration(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system configuration:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["configuration"].Children))
-		return nil
-	}
-
-	if args[0] != "rescue" {
-		return fmt.Errorf("unknown request system configuration command: %s", args[0])
-	}
-
-	if len(args) < 2 {
-		fmt.Println("request system configuration rescue:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["configuration"].Children["rescue"].Children))
-		return nil
-	}
-
-	switch args[1] {
-	case "save":
-		if err := c.store.SaveRescueConfig(); err != nil {
-			return err
-		}
-		fmt.Println("Rescue configuration saved")
-		return nil
-
-	case "delete":
-		if err := c.store.DeleteRescueConfig(); err != nil {
-			return err
-		}
-		fmt.Println("Rescue configuration deleted")
-		return nil
-
-	default:
-		return fmt.Errorf("unknown request system configuration rescue command: %s", args[1])
-	}
-}
-
-func (c *CLI) handleRequestSecurity(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request security:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["security"].Children))
-		return nil
-	}
-	switch args[0] {
-	case "ipsec":
-		if len(args) < 3 || args[1] != "sa" || args[2] != "clear" {
-			fmt.Println("request security ipsec sa:")
-			writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["security"].Children["ipsec"].Children["sa"].Children))
-			return nil
-		}
-		if c.ipsec == nil {
-			return fmt.Errorf("IPsec manager not available")
-		}
-		count, err := c.ipsec.TerminateAllSAs()
-		if err != nil {
-			return err
-		}
-		fmt.Printf("Cleared %d IPsec SA(s)\n", count)
-		return nil
-	case "wireguard":
-		return c.handleRequestSecurityWireguard(args[1:])
-	case "policies":
-		return c.handleRequestSecurityPolicies(args[1:])
-	default:
-		return fmt.Errorf("unknown request security target: %s", args[0])
-	}
-}
-
-// handleRequestSecurityPolicies implements `request security policies check`
-// (fable-167 C-1c, #4314): a static analysis of the configured policy set
-// that reports shadowed / redundant rules. Junos `request security policies
-// check` validates the compiled policy DB; xpf runs a conservative
-// name-set-containment pass over config.ZonePairPolicies (no dataplane call —
-// this is a config lint). It never mutates config.
-func (c *CLI) handleRequestSecurityPolicies(args []string) error {
-	if len(args) == 0 || args[0] != "check" {
-		fmt.Println("request security policies:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["security"].Children["policies"].Children))
-		return nil
-	}
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("no active configuration")
-		return nil
-	}
-	findings := analyzePolicyShadowing(cfg)
-	if len(findings) == 0 {
-		fmt.Println("Policy check complete: no shadowed or redundant policies detected.")
-		return nil
-	}
-	fmt.Printf("Policy check complete: %d issue(s) detected.\n\n", len(findings))
-	for _, f := range findings {
-		fmt.Println(f)
-	}
-	return nil
-}
-
-// handleRequestSecurityWireguard implements `request security wireguard
-// generate-private-key` (#1434 Increment 1): a stateless utility that
-// prints a fresh WireGuard key pair for the operator to paste into
-// config. Junos `request` semantics — it does NOT mutate config and
-// needs no dataplane (pure-Go X25519 keygen). Mirrors `wg genkey` /
-// `wg pubkey` output in WireGuard-canonical base64.
-func (c *CLI) handleRequestSecurityWireguard(args []string) error {
-	if len(args) == 0 || args[0] != "generate-private-key" {
-		fmt.Println("request security wireguard:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["security"].Children["wireguard"].Children))
-		return nil
-	}
-	kp, err := wgkey.Generate()
-	if err != nil {
-		return fmt.Errorf("generate WireGuard key: %w", err)
-	}
-	fmt.Printf("Private key: %s\n", kp.PrivateKey)
-	fmt.Printf("Public key:  %s\n", kp.PublicKey)
-	return nil
 }
